refactor: split edit_job_grid_logic.js into smaller modules
This commit refactors the monolithic `edit_job_grid_logic.js` file into smaller, more manageable modules. This improves code organization, readability, and maintainability.

The following modules were created:

- `job/client_lookup.js`: Handles client lookup functionality.
- `job/deseralise_job_pricing.js`: Deserializes job pricing data.
- `job/deserialize_job_files.js`: Deserializes job files data.
- `job/edit_job_buttons.js`: Handles button click events on the edit job form.
- `job/edit_job_form_autosave.js`: Implements autosave functionality for the edit job form.
- `job/edit_job_grid_logic.js`: Main entry point for grid logic, responsible for initializing and managing grids.
- `job/grid/columns.js`: Defines grid column definitions.
- `job/grid/grid_initialization.js`: Handles grid initialization and management.
- `job/grid/grid_options.js`: Creates grid options for different grid types.
- `job/grid/grid_utils.js`: Contains utility functions for grid calculations and operations.
- `job/grid/parsers.js`: Contains number and currency parsers.
- `job/job_buttons/button_handlers.js`: Handles button click events.
- `job/job_details_loader.js`: Loads job details.

This change improves the overall structure of the codebase and makes it easier to understand and modify the grid logic.

feat: add job grid functionality

This commit introduces a comprehensive set of features related to job grids, including:

- Grid utilities for calculations and event handling.
- Parsers for formatting and data conversion.
- Grid options for revenue and cost tables.
- Button handlers for various job actions.
- Utility functions for button interactions and URL manipulation.
- Email handling logic for sending quotes.
- Job event management for adding events to the timeline.
- Modal handlers for displaying instructions and previews.
- Xero document management for creating and deleting documents.
- Job details loading for expanding text areas.
- Job file handling for uploading, deleting, and managing files.

These changes enhance the job management capabilities of the application by providing a structured and interactive grid system for pricing, revenue, and cost calculations, as well as improved file and document management.
