+++ workflow/templates/jobs/edit_job_ajax.html
@@ -211,10 +211,10 @@
     style="display:none;">{{ latest_job_pricings_json|safe }}</script>
 <script type="application/json" id="historicalJobPricingsData"
     style="display:none;">{{ historical_job_pricings_json|safe }}</script>
-<script type="module" src="{% static 'js/deseralise_job_pricing.js' %}"></script>
-<script type="module" src="{% static 'js/edit_job_form_autosave.js' %}"></script>
-<script type="module" src="{% static 'js/job_file_handling.js' %}"></script>
-<script type="module" src="{% static 'js/client_lookup.js' %}"></script>
-<script type="module" src="{% static 'js/edit_job_grid_logic.js' %}"></script>
+<script type="module" src="{% static 'js/job/deseralise_job_pricing.js' %}"></script>
+<script type="module" src="{% static 'js/job/edit_job_form_autosave.js' %}"></script>
+<script type="module" src="{% static 'js/job/job_file_handling.js' %}"></script>
+<script type="module" src="{% static 'js/job/client_lookup.js' %}"></script>
+<script type="module" src="{% static 'js/job/edit_job_grid_logic.js' %}"></script>
 <script type="module" src="{% static 'js/time_conversion.js' %}"></script>
 {% endblock %}
